Extract config utils to separate module
