+++ extract_features.py
@@ -15,7 +15,8 @@ from winnow.feature_extraction.model import default_model_path
 from winnow.storage.db_result_storage import DBResultStorage
 from winnow.storage.repr_storage import ReprStorage
 from winnow.storage.repr_utils import bulk_read, bulk_write
-from winnow.utils import scan_videos, create_video_list, scan_videos_from_txt, resolve_config, reprkey_resolver
+from winnow.utils import scan_videos, create_video_list, scan_videos_from_txt, reprkey_resolver
+from winnow.utils.config import resolve_config
 
 logging.getLogger().setLevel(logging.ERROR)
 logging.getLogger("winnow").setLevel(logging.INFO)
